perf: PlanningViewer
